+++ scraper/src/alerting/index.ts
@@ -1,4 +1,4 @@
-export * from './alertNeeded';
-export * from './triggerAlertsForAvailable';
-export * from './triggerAlertsForDiscount';
-export * from './triggerAlertsForNewItem';
+export * from './alertNeeded.js';
+export * from './triggerAlertsForAvailable.js';
+export * from './triggerAlertsForDiscount.js';
+export * from './triggerAlertsForNewItem.js';
