+++ scraper/src/index.ts
@@ -1,12 +1,11 @@
 import { Clog, LOGLEVEL } from '@fdebijl/clog';
 import { Price, listItems, insertPrice, upsertItem, } from 'wtcheap.shared';
 
-import { getCurrentItems, deepCheckItem } from './scrapers';
-import { availableAlertNeeded, discountAlertNeeded, triggerAlertsForAvailable, triggerAlertsForDiscount, triggerAlertsForItems } from './alerting';
-import { Price } from './domain';
-import { SHOP_2022_SELECTORS, TARGET_ROOTS } from './constants';
-import { waybackMain } from './wayback';
-import { ensureIndices } from './db/ensureIndices';
+import { getCurrentItems, deepCheckItem } from './scrapers/index.js';
+import { availableAlertNeeded, discountAlertNeeded, triggerAlertsForAvailable, triggerAlertsForDiscount, triggerAlertsForItems } from './alerting/index.js';
+import { SHOP_2022_SELECTORS, TARGET_ROOTS } from './constants.js';
+import { waybackMain } from './wayback.js';
+import { ensureIndices } from './db/ensureIndices.js';
 
 export const clog = new Clog(LOGLEVEL.DEBUG);
