+++ scraper/src/scrapers/index.ts
@@ -1,7 +1,7 @@
-export * from './deepCheckItem';
-export * from './getCurrentItems';
-export * from './isItemBuyable';
-export * from './matchSelectors';
-export * from './getDetailsOnPage';
-export * from './getItemsOnPage';
-export * from './findNon404Memento';
+export * from './deepCheckItem.js';
+export * from './getCurrentItems.js';
+export * from './isItemBuyable.js';
+export * from './matchSelectors.js';
+export * from './getDetailsOnPage.js';
+export * from './getItemsOnPage.js';
+export * from './findNon404Memento.js';
