+++ shared/package.json
@@ -2,6 +2,7 @@
   "name": "wtcheap.shared",
   "private": true,
   "license": "MIT",
+  "type": "module",
   "version": "1.0.0",
   "main": "dist/index.js",
   "scripts": {
