+++ shared/src/db/alerts/index.ts
@@ -1,4 +1,4 @@
-export * from './delete';
-export * from './exists';
-export * from './find';
-export * from './insert';
+export * from './delete.js';
+export * from './exists.js';
+export * from './find.js';
+export * from './insert.js';
