+++ shared/src/db/connect.ts
@@ -1,6 +1,6 @@
 import { Db, MongoClient } from 'mongodb';
 
-import { MONGODB_URI } from '../constants';
+import { MONGODB_URI } from '../constants.js';
 
 export const connect = async (): Promise<Db> => {
   const client = new MongoClient(MONGODB_URI);
