+++ shared/src/db/index.ts
@@ -1,5 +1,5 @@
-export * from './connect';
+export * from './connect.js';
 
-export * from  './alerts';
-export * from  './items';
-export * from  './prices';
+export * from  './alerts/index.js';
+export * from  './items/index.js';
+export * from  './prices/index.js';
