+++ shared/src/db/items/upsert.ts
@@ -1,8 +1,8 @@
 import { LOGLEVEL } from '@fdebijl/clog';
 
-import { Item } from '../../domain';
-import { connect } from '../connect';
-import { clog } from '../../index';
+import { Item } from '../../domain/index.js';
+import { connect } from '../connect.js';
+import { clog } from '../../index.js';
 
 export const upsertItem = async (item: Item): Promise<void> => {
   const db = await connect();
