+++ shared/src/db/prices/index.ts
@@ -1,2 +1,2 @@
-export * from './insert';
-export * from './forItem';
+export * from './insert.js';
+export * from './forItem.js';
