+++ shared/src/db/prices/insert.ts
@@ -1,5 +1,5 @@
-import { connect } from '../connect';
-import { Price } from '../../domain';
+import { connect } from '../connect.js';
+import { Price } from '../../domain/index.js';
 
 export const insertPrice = async (price: Price): Promise<void> => {
   const db = await connect();
