+++ shared/src/domain/index.ts
@@ -1,7 +1,7 @@
-export * from './alert';
-export * from './item';
-export * from './mail';
-export * from './mailfactory';
-export * from './price';
-export * from './memento';
-export * from './selectorset';
+export * from './alert.js';
+export * from './item.js';
+export * from './mail.js';
+export * from './mailfactory.js';
+export * from './price.js';
+export * from './memento.js';
+export * from './selectorset.js';
