+++ shared/src/mailfactory/index.ts
@@ -1,4 +1,4 @@
-export * from './mailAvailableFactory';
-export * from './mailDiscountFactory';
-export * from './mailNewItemFactory';
-export * from './mailTokenFactory';
+export * from './mailAvailableFactory.js';
+export * from './mailDiscountFactory.js';
+export * from './mailNewItemFactory.js';
+export * from './mailTokenFactory.js';
