+++ tsconfig.json
@@ -1,11 +1,11 @@
 {
   "compilerOptions": {
     "target": "es2020",
-    "module": "Node16",
+    "module": "NodeNext",
     "declaration": true,
     "sourceMap": true,
     "strict": true,
-    "moduleResolution": "node16",
+    "moduleResolution": "nodenext",
     "esModuleInterop": true,
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true
